ExamenAñoPasado y P1e1

+++ ExamenAnoPasado/src/Modulo/Producto.java
@@ -3,13 +3,14 @@ package Modulo;
 public class Producto {
     //Atributos
     private String nombre;
-    private Double precio;
+    private Float precio;
     private int Stock;
 
     //Constructor
     public Producto() {
     }
-    public Producto(String nombre, Double precio, int stock) {
+
+    public Producto(String nombre, Float precio, int stock) {
         this.nombre = nombre;
         this.precio = precio;
         Stock = stock;
@@ -19,8 +20,8 @@ public class Producto {
     public String getNombre() {return nombre;}
     public void setNombre(String nombre) {this.nombre = nombre;}
 
-    public Double getPrecio() {return precio;}
-    public void setPrecio(Double precio) {this.precio = precio;}
+    public Float getPrecio() {return precio;}
+    public void setPrecio(Float precio) {this.precio = precio;}
 
     public int getStock() {return Stock;}
     public void setStock(int stock) {Stock = stock;}
